Move dashboard components into sub folder
