+++ packages/frontend/src/pages/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import DashboardMyBugsCard from '../components/DashboardMyBugsCard';
-import DashboardProjectsCard from '../components/DashboardProjectsCard';
+import DashboardMyBugsCard from '../components/dashboard/DashboardMyBugsCard';
+import DashboardProjectsCard from '../components/dashboard/DashboardProjectsCard';
 
 const Dashboard = () =>
